Moved import card logic outside of the MainActivity
