+++ app/src/main/java/com/github/nacabaro/vbhelper/screens/settingsScreen/SettingsScreenController.kt
@@ -4,4 +4,5 @@ interface SettingsScreenController {
     fun onClickOpenDirectory()
     fun onClickImportDatabase()
     fun onClickImportApk()
+    fun onClickImportCard()
 }
